requirements.txt
anytree==2.8.0
GitPython==3.1.3